requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
supabase==2.10.0
httpx==0.27.2
openai==1.59.7
orjson==3.10.12
